# Conflicts:
#	routes/web.php

+++ resources/views/users/index.blade.php
@@ -6,8 +6,10 @@
 @section('content')
 
     <div class="col-lg-10 col-lg-offset-1">
-        <h1><i class="fa fa-users"></i> Пользователи <a href="{{ route('roles.index') }}" class="btn btn-default pull-right">Роли</a>
-            <a href="{{ route('permissions.index') }}" class="btn btn-default pull-right">Операции</a></h1>
+        <h1><i class="fa fa-users"></i> Пользователи
+            {{--<a href="{{ route('roles.index') }}" class="btn btn-default pull-right">Роли</a>--}}
+            {{--<a href="{{ route('permissions.index') }}" class="btn btn-default pull-right">Операции</a>--}}
+        </h1>
         <hr>
         <div class="table-responsive">
             <table class="table table-bordered table-striped">
@@ -20,7 +22,7 @@
                     <th>Телефон</th>
                     <th>Email</th>
                     <th>Дата/Время добавления</th>
-                    <th>Роль пользователя</th>
+                    <th>Роль</th>
                     <th>Действия</th>
                 </tr>
                 </thead>
@@ -37,21 +39,22 @@
                         <td>{{ $user->created_at }}</td>
                         <td>{{ $user->roles()->pluck('name')->implode(' ') }}</td>{{-- Retrieve array of roles associated to a user and convert to string --}}
                         <td>
-                            <p><a href="{{ route('users.edit', $user->id) }}" class="btn btn-info pull-left" style="margin-right: 3px;">Редактировать</a></p>
+                            @if($user->hasRole('Менеджер'))
+                                <p><a href="{{ route('users.edit', $user->id) }}" class="btn btn-info pull-left" style="margin-right: 3px;">Редактировать</a></p>
 
-                            {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id] ]) !!}
-                            {!! Form::submit('Удалить', ['class' => 'btn btn-danger']) !!}
-                            {!! Form::close() !!}
+                                {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id] ]) !!}
+                                {!! Form::submit('Удалить', ['class' => 'btn btn-danger']) !!}
+                                {!! Form::close() !!}
+                            @endif
 
                         </td>
                     </tr>
                 @endforeach
                 </tbody>
-
             </table>
         </div>
 
-        {{--<a href="{{ route('users.create') }}" class="btn btn-success">Добавить нового пользователя</a>--}}
+        <p><a href="{{ route('invitation-key') }}" class="btn btn-success"> Пригласительный ключ для регистрации </a></p>
 
     </div>
